Manejo de restricciones segun el rol de usuario

+++ TPI-Ecommerce/Controllers/ClientController.cs
@@ -3,6 +3,7 @@ using Application.Models;
 using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Mvc;
+using System.Security.Claims;
 
 namespace TPI_Ecommerce.Controllers
 {
@@ -15,48 +16,87 @@ namespace TPI_Ecommerce.Controllers
         {
             _service = clientService;
         }
+
+        private bool IsUserInRol(string rol)
+        {
+            var claimsRol = User.Claims.FirstOrDefault(u => u.Type == ClaimTypes.Role);
+            return claimsRol is not null && claimsRol.Value == rol;
+        }
+
         [HttpGet]
         [Authorize]
         public IActionResult GetAll()
         {
-            return Ok(_service.GetAll());
+            if(IsUserInRol("Admin"))
+            {
+                return Ok(_service.GetAll());
+            }
+            return Forbid();
+          
         }
 
         [HttpGet("{id}")]
         [Authorize]
         public IActionResult GetById(int id)
         {
-            var client = _service.Get(id);
-            if (client == null)
+
+           if(IsUserInRol("Admin"))
             {
-                NotFound("No se encontro el cliente");
+                var client = _service.Get(id);
+                if (client == null)
+                {
+                    NotFound("No se encontro el cliente");
+                }
+
+                return Ok(client);
             }
 
-            return Ok(client);
+            return Forbid();
         }
 
 
         [HttpPost]
-      
         public IActionResult AddClient([FromBody] ClientCreateDto client)
         {
             _service.Add(client);
             return Ok("Client added succesfully");
-
         }
+
+        [HttpPut("{id}")]
+        [Authorize]
+        public IActionResult UpdateClient([FromRoute] int id, [FromBody] ClientUpdateDto dto)
+        {
+            var userUpdate = _service.Get(id);
+            if(userUpdate == null)
+            {
+                return NotFound($"No se encontro el cliente con el ID: {id}");
+            }
+
+            if(IsUserInRol("Admin"))
+            {
+                _service.Update(id, dto);
+                return Ok("El cliente fue actualizado correctamente");
+            }
+            return Forbid();
+        }
+
 
         [HttpDelete("{id}")]
         [Authorize]
         public IActionResult DeleteClient(int id)
         {
-            var client = _service.Get(id);
-            if (client == null)
+           if(IsUserInRol("Admin"))
             {
-                return NotFound("No se encontro el cliente");
-            }
+                var client = _service.Get(id);
+                if (client == null)
+                {
+                    return NotFound("No se encontro el cliente");
+                }
 
-            _service.Delete(id);
-            return Ok("El cliente fue eliminado");
+                _service.Delete(id);
+                return Ok("El cliente fue eliminado");
+            }
+            return Forbid();
         }
 
     }
